+++ src/trafficgen/encounter.py
@@ -19,6 +19,7 @@ from trafficgen.types import (
     EncounterType,
     GeoPosition,
     Initial,
+    Leg,
     OwnShip,
     ShipStatic,
     SituationInput,
@@ -31,6 +32,7 @@ from trafficgen.utils import (
     calculate_position_at_certain_time,
     convert_angle_0_to_2_pi_to_minus_pi_to_pi,
     convert_angle_minus_pi_to_pi_to_0_to_2_pi,
+    deg_2_rad,
 )
 
 
@@ -185,7 +187,7 @@ def generate_encounter(
                     lat_lon0,
                     target_ship_sog,
                     target_ship_cog,
-                    target_ship_position_future,
+                    target_ship_initial_position,
                     desired_encounter_type,
                     settings,
                 )
@@ -213,8 +215,9 @@ def generate_encounter(
             heading=target_ship_cog,
             nav_status=AisNavStatus.UNDER_WAY_USING_ENGINE,
         )
+        leg: Leg = Leg(sog=target_ship_sog)
         target_ship_waypoint0 = Waypoint(
-            position=target_ship_initial_position.model_copy(deep=True), turn_radius=None, leg=None
+            position=target_ship_initial_position.model_copy(deep=True), turn_radius=None, leg=leg
         )
 
         future_position_target_ship = calculate_position_at_certain_time(
@@ -225,7 +228,9 @@ def generate_encounter(
             settings.situation_length,
         )
 
-        target_ship_waypoint1 = Waypoint(position=future_position_target_ship, turn_radius=None, leg=None)
+        target_ship_waypoint1 = Waypoint(
+            position=future_position_target_ship, turn_radius=None, leg=leg.model_copy(deep=True)
+        )
         waypoints = [target_ship_waypoint0, target_ship_waypoint1]
 
         target_ship = TargetShip(static=target_ship_static, initial=target_ship_initial, waypoints=waypoints)
@@ -326,6 +331,8 @@ def define_own_ship(
     own_ship_static: ShipStatic,
     encounter_settings: EncounterSettings,
     lat_lon0: GeoPosition,
+    *,
+    overwrite_ownship_initial_coord: bool = False,
 ) -> OwnShip:
     """
     Define own ship based on information in desired traffic situation.
@@ -340,6 +347,8 @@ def define_own_ship(
         Necessary setting for the encounter
     lat_lon0 : GeoPosition
         Reference point, latitudinal [rad] and longitudinal [rad]
+    overwrite_ownship_initial_coord : bool
+        If True, the ownship initial coordinate from SituationInput is overwritten with the lat_lon0 value.
 
     Returns
     -------
@@ -347,12 +356,18 @@ def define_own_ship(
         Own ship including static, initial and waypoints.
     """
     own_ship_initial: Initial = desired_traffic_situation.own_ship.initial
+
+    if overwrite_ownship_initial_coord:
+        # assign lat_lon0, but convert them to radians first!
+        own_ship_initial.position = GeoPosition(lat=deg_2_rad(lat_lon0.lat), lon=deg_2_rad(lat_lon0.lon))
+
     own_ship_waypoints: list[Waypoint] = []
+    leg: Leg = Leg(sog=own_ship_initial.sog)
     if desired_traffic_situation.own_ship.waypoints is None:
         # If waypoints are not given, let initial position be the first waypoint,
         # then calculate second waypoint some time in the future
         own_ship_waypoint0 = Waypoint(
-            position=own_ship_initial.position.model_copy(deep=True), turn_radius=None, leg=None
+            position=own_ship_initial.position.model_copy(deep=True), turn_radius=None, leg=leg
         )
         ship_position_future = calculate_position_at_certain_time(
             own_ship_initial.position,
@@ -361,14 +376,14 @@ def define_own_ship(
             own_ship_initial.cog,
             encounter_settings.situation_length,
         )
-        own_ship_waypoint1 = Waypoint(position=ship_position_future, turn_radius=None, leg=None)
+        own_ship_waypoint1 = Waypoint(position=ship_position_future, turn_radius=None, leg=leg.model_copy(deep=True))
         own_ship_waypoints = [own_ship_waypoint0, own_ship_waypoint1]
     elif len(desired_traffic_situation.own_ship.waypoints) == 1:
         # If one waypoint is given, use initial position as first waypoint
         own_ship_waypoint0 = Waypoint(
-            position=own_ship_initial.position.model_copy(deep=True), turn_radius=None, leg=None
+            position=own_ship_initial.position.model_copy(deep=True), turn_radius=None, leg=leg
         )
-        own_ship_waypoint1 = desired_traffic_situation.own_ship.waypoints[0]
+        own_ship_waypoint1 = desired_traffic_situation.own_ship.waypoints[0].model_copy(deep=True)
         own_ship_waypoints = [own_ship_waypoint0, own_ship_waypoint1]
     else:
         own_ship_waypoints = desired_traffic_situation.own_ship.waypoints