+++ src/trafficgen/write_traffic_situation_to_file.py
@@ -101,11 +101,13 @@ def convert_ship_data_from_si_units_to_maritime(ship: T_ship) -> T_ship:
                     waypoint.leg.starboard_xtd = round(m_2_nm(waypoint.leg.starboard_xtd), 2)
                 if waypoint.leg.portside_xtd is not None:
                     waypoint.leg.portside_xtd = round(m_2_nm(waypoint.leg.portside_xtd), 2)
+                if waypoint.leg.sog is not None:
+                    waypoint.leg.sog = round(m_pr_s_2_knot(waypoint.leg.sog), 1)
                 if waypoint.leg.data is not None and waypoint.leg.data.sog is not None:
                     assert waypoint.leg.data.sog.value is not None
                     assert waypoint.leg.data.sog.interp_start is not None
                     assert waypoint.leg.data.sog.interp_end is not None
-                    waypoint.leg.data.sog.value = round(m_pr_s_2_knot(waypoint.leg.data.sog.value), 2)
+                    waypoint.leg.data.sog.value = round(m_pr_s_2_knot(waypoint.leg.data.sog.value), 1)
                     waypoint.leg.data.sog.interp_start = round(m_2_nm(waypoint.leg.data.sog.interp_start), 2)
                     waypoint.leg.data.sog.interp_end = round(m_2_nm(waypoint.leg.data.sog.interp_end), 2)