+++ CHANGELOG.md
@@ -5,10 +5,26 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 
 ## [Unreleased]
 
+### Changed
+
+## [0.8.4] - 2025-11-26
+
+### Changed
+
+* Added sog to each waypoint leg. Updated documentation. Generated new baseline_situations_generated files. Updated types.
+* Fixed bug where future position of target ship was used instead of initial position of target ship to check encounter evolvement. Updated documentation to make this functionality clearer.
+* Added option to pass in a single input situation .json file, not only a folder.
+
 ## [0.8.3] - 2025-10-31
+
+### Changed
+
 * [bug] Fixed sourcery dependency constraint to avoid platform compatibility issues on Linux by constraining to <1.41.
 
 ## [0.8.2] - 2025-10-03
+
+### Changed
+
 * [bug] fixed bug that all target ships had id = 10.
 * [bug] fixed duplicate -v parameter warning.
 * If own ship ID is not given in own_ship.json file, then ID is set to 1. Target ship IDs will follow from own ship ID.
@@ -16,11 +32,15 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 * Updated documentaion of target ships in input_files.rst
 
 ## [0.8.1] - 2025-05-12
+
+### Changed
+
 * Changed first assert in  test_basic_cli() from 0 to 2.
 
 ## [0.8.0] - 2025-05-09
 
 ### Changed
+
 * vectorTime may now be a single float (exact time) or a list of two floats [from, to]. vectorRange is removed from settings file.
 * evolveTime is changed to situationDevelopTime.
 * Removed Basemap-package and added pyproj package.